# Pandas Mapper Performance Refactor