@C-SELLERS 2021-09-07
After testing it has been decided this is a unnecessary change and has a negative affect on our most practical case of using dataframes. See test results in PR description.